변경된 report type이 정상적으로 출력되도록 ui 수정
