+++ src/util/createReport.ts
@@ -1,6 +1,6 @@
 import { ProfileDefinition, StatReport } from "@/types/report";
 
-export function createReport(type: StatReport["type"], name: string, defs: ProfileDefinition[]): StatReport {
+export function createDefaultReport(type: StatReport["type"], name: string, defs: ProfileDefinition[]): StatReport {
   const base = { name, createdAt: new Date().toISOString(), profileDefinitions: defs };
   return type === "performance"
       ? ({ ...base, type, report: { statDefinitions: [], performanceRecords: [] } })
